+++ public/app/plugins/datasource/influxdb/influx_query.ts
@@ -194,7 +194,7 @@ export default class InfluxQuery {
     }
 
     var escapedValues = _.map(value, kbn.regexEscape);
-    return escapedValues.join('|');
+    return '(' + escapedValues.join('|') + ')';
   }
 
   render(interpolate?) {